app/tools/interpreter_agent/requirements.txt
fastapi
uvicorn
openai-agents
pydantic
requests
python-dotenv
